backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
aiofiles==23.2.1
pydantic==2.6.1
pydantic-settings==2.1.0
numpy==1.26.4
Pillow==10.2.0
scikit-learn==1.4.0
faiss-cpu==1.7.4
selenium==4.17.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
requests==2.31.0
librosa==0.10.1
soundfile==0.12.1
tensorflow==2.15.0
openvino==2023.3.0
opencv-python-headless==4.9.0.80
python-magic==0.4.27
httpx==0.27.0
